requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
pydantic
langchain
langchain-core
langgraph
langgraph-supervisor
google-cloud-aiplatform
vertexai
google-cloud-texttospeech
